requirements.txt
sanic
tortoise-orm
aiomysql
PyJWT
aioredis
pyotp
